+++ PoshToolbox/Public/Use-NullCoalescing.ps1
@@ -40,6 +40,3 @@ function Use-NullCoalescing {
         }
     }
 }
-
-
-Set-Alias -Name ?? -Value Use-NullCoalescing